organised files in different folders, app runs
